test: capability の露出検査を builtins のテストに寄せる (#876)
別ファイルに置くと builtins のモジュールグラフを二重に collect することに
なり、全体実行の負荷でメディアプロキシのテストがタイムアウトしていた。
同じ import を持つ既存テストに同居させる。

Refs #876
